Extend syntax for ALTER TEXT SEARCH CONFIGURATION ALTER/ADD.
Now operators AND, OR and THEN are valid in text search configuration
of the dictionaries pipeline as well as parentheses and stored in pg_ts_config_map.
Currently all operators have same priority (to be changed in the future).
The comma sign is equal to the OR operator, so old sytax is supported.
The internal logic of the text search engine is not modified yet, so all
additional information is stored but not used by the postgres.