Introducing de-TOAST iterator for Generic Toaster
Detoasting is performed as an iterator - part by part, when full value
detoast is not required, as for textpos function. Iterator is introduced
as new structure DetoastIterator and iterator creation and next functions
are available with TOAST API virtual functions table.
Detoast iterator is not a part of TOAST API but is a revival of patch in [2].
Iterators are very useful for cases when full detoasting is not required -
like searching in TOASTed data.
Detoast iterator implementation is hidded inside Generic Toaster, and
provides two additional functions to user:
detoast_iterator_create returning DetoastIterator as Datum, to hide internal
datails from user;
detoast_iterate_next receiving DetoastIterator and buffer to fill as
parameters.

Author: Teodor Sigaev <teodor@sigaev.ru>
Author: Oleg Bartunov <obartunov@postgrespro.ru>
Author: Nikita Glukhov <n.gluhov@postgrespro.ru>
Author: Nikita Malakhov <n.malakhov@postgrespro.ru>

[1] Discussion: https://www.postgresql.org/message-id/flat/224711f9-83b7-a307-b17f-4457ab73aa0a@sigaev.ru
[2] Discussion: https://www.postgresql.org/message-id/flat/CAL-OGktjn9wHLHhEem58qJXMBUAVWa8-VrDoCmceAxXGTUM_BA%40mail.gmail.com#17cd4f49591c2ea3cba364c13e129147

Detoast Iterator code cleanup and fix