[WIP] 64-bit TOAST Value ID
The TOAST mechanics has one implicit limitation that is quite often
encountered in large production databases with high load rates. The
TOAST Value ID is a 4-byte integer, so it has maximum value of
4294967295 which allows maximum of 4294967295 values in single TOAST
table. If TOAST values are frequently updated and autovacuum does
not keep up with the update rate - sonner or later this table will
be out of available values IDs for INSERT/UPDATE commands, and on
INSERT/UPDATE of the TOASTed value in this TOAST table the database
will enter infinite loop of sorting out TOAST values IDs looking
for available one.
This is discuissed in [1] and [2].
This patch is a proposal which introduces new TOAST pointer with
extended Value ID - 8 bytes instead of 4, and an individual value
counter and cache for each TOAST table. This will solve problem
of expending TOAST value ID once and for all, for tables under
32 Tb.
Also, this patch keeps backward compatibility with the old TOAST
pointer format.

This is a WIP patch, so we'll be grateful for any feedback and
proposals on this solution.

[1] https://www.postgresql.org/message-id/flat/16722-93043fb459a41073%40postgresql.org
[2] https://www.postgresql.org/message-id/CAN-LCVPRvRzxeUdYdDCZ7UwZQs1NmZpqBUCd%3D%2BRdMPFTyt-bRQ%40mail.gmail.com