Consider the number of columns when estimating the Sort's cost.
During the multicolumn sort, we apply a comparison operator to each column one
by one. So, even keeping aside the cost of a specific operator, we should take
into account the number of its calls.

This code impacts the cost model of GatherMerge, IncrementalSort, Sort, and
MergeAppend. It also alters the balance between ordered and hashed operations,
which should be benchmarked and tuned carefully.