Remember pathkey's source expression.
To consistently estimate the number of groups using the pathkeys list in the
sorting operation, we need to identify the expression used as a sort key.
Pathkey refers to the equivalence class, which can contain a wide variety of
expressions. To somehow stabilise the expression's choice, save the link to the
origin expression, triggering the PathKey creation. While this approach doesn't
always guarantee the right choice, it makes dependence on the SQL string less
obvious.